+++ package.json
@@ -23,6 +23,7 @@
     "@tanstack/react-query": "^5.18.1",
     "axios": "^1.6.5",
     "axios-observable": "^2.0.0",
+    "immer": "^10.1.1",
     "moment": "^2.30.1",
     "react": "18.3.1",
     "react-freeze": "^1.0.3",
