+++ src/screen/dashboard/HomeTab/Home/components/export.ts
@@ -1,0 +1,2 @@
+export * from './Detail';
+export * from './Avatar';
