+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './Colors';
 export * from './FontStyle';
 export * from './responsive';
 export * from './CustomStyle';
+export * from './date';
