+++ src/zustand/function/Function.ts
@@ -1,0 +1,6 @@
+import { ZustandFnc } from '../export';
+import { useHomeZustand } from '../home/HomeStore';
+
+export const zustandFnc: ZustandFnc = {
+  useHomeZustand,
+};
