+++ src/zustand/home/HomeStore.ts
@@ -1,23 +1,15 @@
 import { create } from 'zustand';
-import { HttpPunchDetailResponse, initial } from '../export';
+import {
+  HomeStoreSetter,
+  HomeStoreState,
+  HttpPunchDetailResponse,
+  initial,
+  SetState,
+} from '../export';
 import { immer } from 'zustand/middleware/immer';
 
-// type OmitFunctions<T> = {
-//   [K in keyof T as T[K] extends (...args: never) => never ? never : K]: T[K];
-// };
-
-export type HomeStoreZustand = {
-  data: HttpPunchDetailResponse;
-};
-
-export type HomeStoreSetter = {
-  setData: (data: HttpPunchDetailResponse) => void;
-};
-
-export type SetState = (fnc: (state: HomeStoreZustand) => void) => void;
-
-export const useGlobalLoad = create<HomeStoreZustand & HomeStoreSetter>()(
-  immer<HomeStoreZustand & HomeStoreSetter>((set: SetState) => ({
+export const useHomeZustand = create<HomeStoreState & HomeStoreSetter>()(
+  immer<HomeStoreState & HomeStoreSetter>((set: SetState<HomeStoreState>) => ({
     data: initial.HomeInitialState,
     setData: (data: HttpPunchDetailResponse) => {
       set(state => {
@@ -27,4 +19,4 @@ export const useGlobalLoad = create<HomeStoreZustand & HomeStoreSetter>()(
   })),
 );
 
-export type useGlobalLoadType = ReturnType<typeof useGlobalLoad>;
+export type UseHomeZustandType = ReturnType<typeof useHomeZustand>;
